backend/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
pydantic==2.9.2
python-dotenv==1.0.1
